Major Refactoring and Chart
