+++ src/components/index.js
@@ -1,0 +1,4 @@
+export { default as SidePanel } from './SidePanel'
+export { default as MetricsChart } from './MetricsChart'
+export { default as ActivitySummary } from './ActivitySummary'
+export { default as TopPanel } from './TopPanel'
