+++ src/data/constants.js
@@ -1,0 +1,26 @@
+export const constants = [
+  {
+    name: 'Gait Time',
+    borders: [0.0, 3.0],
+    healthyArea: [1.69, 1.1],
+    cohortMedian: 1.39,
+  },
+  {
+    name: 'Step Variability',
+    borders: [0.0, 3.5],
+    healthyArea: [0.38, 1.74],
+    cohortMedian: 1.06,
+  },
+  {
+    name: 'Cadence',
+    borders: [40, 150],
+    healthyArea: [70.12, 106.37],
+    cohortMedian: 88.24,
+  },
+  {
+    name: 'Median COP Speed',
+    borders: [0.0, 3.0],
+    healthyArea: [0.86, 1.68],
+    cohortMedian: 1.27,
+  },
+]
