+++ src/index.js
@@ -1,12 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import {
-  BrowserRouter as Router,
-  Route,
-  Switch,
-  Redirect,
-} from 'react-router-dom'
-
+import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom'
 import './style.css'
 import Desktop1 from './views/desktop1'
 import NotFound from './views/not-found'
@@ -15,9 +9,9 @@ const App = () => {
   return (
     <Router>
       <Switch>
-        <Route component={Desktop1} exact path="/" />
-        <Route component={NotFound} path="**" />
-        <Redirect to="**" />
+        <Route component={Desktop1} exact path='/' />
+        <Route component={NotFound} path='**' />
+        <Redirect to='**' />
       </Switch>
     </Router>
   )
