+++ src/utils/index.js
@@ -1,0 +1,3 @@
+import { findMedian } from './helperFunctions'
+
+export { findMedian }
