+++ src/views/desktop1.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
+
 .cadencegraph {
   top: 915.547119140625px;
   left: 210px;
@@ -22,7 +24,7 @@
   position: absolute;
   z-index: 9999; /* Adjust the z-index value as needed */
   width: 500px; /* Adjust the width as needed */
-  height: 200px;
+  height: 20px;
 }
 
 .copspeedgraph {
@@ -68,35 +70,7 @@
   font-stretch: normal;
   text-decoration: none;
 }
-.desktop1-text004 {
-  top: 216.62161254882812px;
-  left: 342.79058837890625px;
-  color: rgba(53, 53, 53, 1);
-  height: auto;
-  position: absolute;
-  font-style: Bold;
-  text-align: left;
-  font-family: Poppins;
-  font-weight: 700;
-  line-height: normal;
-  font-stretch: normal;
-  text-decoration: none;
-}
 
-.desktop1-text008 {
-  top: 216.62161254882812px;
-  left: 342.79058837890625px;
-  color: rgba(53, 53, 53, 1);
-  height: auto;
-  position: absolute;
-  font-style: SemiBold;
-  text-align: left;
-  font-family: Poppins;
-  font-weight: 600;
-  line-height: normal;
-  font-stretch: normal;
-  text-decoration: none;
-}
 .desktop1-text010 {
   top: 216.62161254882812px;
   left: 342.79058837890625px;
@@ -125,9 +99,8 @@
   width: 100%;
   height: 91px;
   position: absolute;
-  background-color:#1db3d8; 
+  background-color: #1db3d8;
 }
-
 
 .desktop1-text012 {
   top: 21.900117874145508px;
@@ -357,7 +330,7 @@
   font-size: 38.9779px;
   font-style: Bold;
   text-align: left;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 700;
   line-height: normal;
   font-stretch: normal;
@@ -379,7 +352,7 @@
   font-size: 11.6934px;
   font-style: Medium;
   text-align: center;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 500;
   line-height: normal;
   font-stretch: normal;
@@ -395,7 +368,7 @@
   font-size: 8.4452px;
   font-style: Medium;
   text-align: center;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 500;
   line-height: normal;
   font-stretch: normal;
@@ -557,7 +530,7 @@
   font-size: 38.9779px;
   font-style: Bold;
   text-align: left;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 700;
   line-height: normal;
   font-stretch: normal;
@@ -595,7 +568,7 @@
   font-size: 8.4452px;
   font-style: Medium;
   text-align: center;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 500;
   line-height: normal;
   font-stretch: normal;
@@ -623,7 +596,7 @@
   font-size: 8.4452px;
   font-style: Medium;
   text-align: center;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 500;
   line-height: normal;
   font-stretch: normal;
@@ -771,7 +744,7 @@
   font-size: 38.9779px;
   font-style: Bold;
   text-align: left;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 700;
   line-height: normal;
   font-stretch: normal;
@@ -796,7 +769,7 @@
   font-size: 11.6934px;
   font-style: Medium;
   text-align: center;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 500;
   line-height: normal;
   font-stretch: normal;
@@ -812,7 +785,7 @@
   font-size: 8.4452px;
   font-style: Medium;
   text-align: center;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 500;
   line-height: normal;
   font-stretch: normal;
@@ -831,7 +804,7 @@
   font-size: 7.79557px;
   font-style: Medium;
   text-align: center;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 500;
   line-height: normal;
   font-stretch: normal;
@@ -960,7 +933,7 @@
   font-size: 38.9779px;
   font-style: Bold;
   text-align: left;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 700;
   line-height: normal;
   font-stretch: normal;
@@ -982,7 +955,7 @@
   font-size: 11.6934px;
   font-style: Medium;
   text-align: center;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 500;
   line-height: normal;
   font-stretch: normal;
@@ -998,7 +971,7 @@
   font-size: 11.6934px;
   font-style: Medium;
   text-align: center;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 500;
   line-height: normal;
   font-stretch: normal;
@@ -1013,7 +986,7 @@
   font-size: 8.4452px;
   font-style: Medium;
   text-align: center;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 500;
   line-height: normal;
   font-stretch: normal;
@@ -1044,7 +1017,7 @@
   font-size: 8.4452px;
   font-style: Medium;
   text-align: center;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 500;
   line-height: normal;
   font-stretch: normal;
@@ -1252,7 +1225,7 @@
   width: 79px;
   height: auto;
   position: absolute;
-  font-size: 10.153660774230957px;
+  font-size: 12px;
   font-style: SemiBold;
   text-align: left;
   font-family: Poppins;
@@ -1268,7 +1241,7 @@
   width: 221px;
   height: auto;
   position: absolute;
-  font-size: 7.3331990242004395px;
+  font-size: 12px;
   font-style: SemiBold;
   text-align: left;
   font-family: Poppins;
@@ -1633,7 +1606,7 @@
   font-size: 8.4452px;
   font-style: Medium;
   text-align: center;
-  font-family: "Poppins";
+  font-family: 'Poppins';
   font-weight: 500;
   line-height: normal;
   font-stretch: normal;
